+++ js/main.js
@@ -1,58 +1,1 @@
-let id = 1;
-let commentId = 1;
-
-const DESCRIPTION = [
-  'Первое',
-  'Второе',
-  'Третье',
-  'Последнее'
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Атос',
-  'Портос',
-  'Арамис',
-  'Милледи'
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomComment = () => {
-  const result = {
-    id: commentId,
-    avatar: `img/avatar-${commentId}.svg`,
-    message: MESSAGES[getRandomInteger(0, MESSAGES.length - 1)],
-    name: NAMES[getRandomInteger(0, NAMES.length - 1)]
-  };
-  commentId += 1;
-  return result;
-};
-
-
-const getArray = () => {
-  const result = {
-    id: id,
-    url: `photos/${id}.jpg`,
-    description: DESCRIPTION[getRandomInteger(0, DESCRIPTION.length - 1)],
-    likes: getRandomInteger(15, 200),
-    comments: Array.from({length: getRandomInteger(1, 3)}, getRandomComment)
-  };
-  id += 1;
-  return result;
-};
-
-const getObjects = Array.from({length: 3}, getArray);
+import {getObjects} from './get-objects';
